Create Choose and Swap
You are given a string s of lower case English alphabets. You can choose any two characters in the string and replace all the occurrences of the first character with the second character and replace all the occurrences of the second character with the first character. Your aim is to find the lexicographically smallest string that can be obtained by doing this operation at most once.